feat!: Deploy multiple Richie instances
This is a major change from the original design of this Tutor plugin.
From running only a Richie instance to allow running multiple Richie
instances. The idea is to use this to run all or some of the Richie
sites inside of a site factory.
Nevertheless there are some configurations that need to be kind of
duplicated because the Open edX need to be know which site is going to
be its marketing site.

fccn/nau-technical#414